requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.7.0
blinker==1.6.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
chromadb==0.3.22
click==8.1.3
clickhouse-connect==0.5.24
cryptography==41.0.2
dataclasses-json==0.6.1
deprecation==2.1.0
dnspython==2.4.2
duckdb==0.7.1
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.95.1
filelock==3.12.0
firebase-admin==6.2.0
Flask==2.2.2
Flask-Cors==4.0.0
frozenlist==1.4.0
fsspec==2023.5.0
google-api-core==2.11.1
google-api-python-client==2.102.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gotrue==1.1.1
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==20.1.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.308
langsmith==0.0.42
lz4==4.3.2
MarkupSafe==2.1.2
marshmallow==3.20.1
marshmallow-enum==1.5.1
midi2audio==0.1.1
mido==1.3.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
openapi-schema-pydantic==1.2.4
orjson==3.9.7
packaging==23.2
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
pip-review==1.3.0
platformdirs==3.9.1
pluggy==1.2.0
postgrest==0.11.0
posthog==3.0.1
pretty-midi==0.2.10
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.0
pydantic_core==2.10.1
PyGuitarPro==0.9.3
PyJWT==2.8.0
pyparsing==3.1.0
pytest==7.4.0
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.4
pytz==2023.3
PyYAML==6.0
realtime==1.0.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.26.1
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.29.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
Werkzeug==2.3.4
yarl==1.9.2
zstandard==0.21.0
